+++ icegraph/__init__.py
@@ -1,11 +1,9 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from . import cache
 from . import console
-from . import extractor
-from . import converter
-from . import dataset
+from .data import extractor, cache, converter
+from . import data
 from . import config
 from . import render
 
@@ -14,7 +12,7 @@ __all__ = [
     "console",
     "extractor",
     "converter",
-    "dataset",
+    "data",
     "config",
     "render",
 ]