+++ icegraph/data/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
+# Developed by Taylor St Jean
+
+from .models import TrainingDataset, ValidationDataset, TestDataset
+from .registry import DatasetRegistry
+
+__all__ = ["TrainingDataset", "ValidationDataset", "TestDataset", "DatasetRegistry"]