+++ icegraph/geometry/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import FeaturePlot
+from .models import Detector